app/requirements.txt
argon2
python-docx
